requirements.txt
Django==5.2.7
djangorestframework
django-filter
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise[brotli]
gunicorn
uvicorn[standard]
# django-prometheus   # décommente si tu veux les métriques Prometheus